Update: 로그인이 안 되면 사인인으로 이동

+++ PeerQA_Web_Django/qa_app/views.py
@@ -20,7 +20,7 @@ def signout(request):
 
         if request.user.is_authenticated:
             logout(request)
-            return HttpResponse(status=204)
+            return redirect("/question/signin/")
         else:
             return HttpResponse(status=401)
 
@@ -28,41 +28,56 @@ def signout(request):
         return HttpResponseNotAllowed(['GET'])
 
 def question_list(request):
-    context = {"question_list": Question.objects.all()}
-    return render(request, "view/question_list.html", context)
+    if request.user.is_authenticated:
+        context = {"question_list": Question.objects.all()}
+        return render(request, "view/question_list.html", context)
+    else:
+        return redirect("/question/signin/")
 
 
 def question_form(request):
-    if request.method == "GET":
-        form = QuestionForm()
-        return render(request, "view/question_form.html", {"form": form, "lectures": lectures})
-    elif request.method == "POST":
-        form = QuestionForm(request.POST)
-        if form.is_valid():
-            form.save()
-        return redirect("/question/")
+    if request.user.is_authenticated:
+        if request.method == "GET":
+            form = QuestionForm()
+            return render(request, "view/question_form.html", {"form": form, "lectures": lectures})
+        elif request.method == "POST":
+            form = QuestionForm(request.POST)
+            if form.is_valid():
+                form.save()
+            return redirect("/question/")
+    else:
+        return redirect("/question/signin/")
 
 def question_detail(request, id):
-    question = Question.objects.get(pk=id)
-    
-    file_name = "/Page" + str(question.lecture_slide) + ".jpeg"
-    src_text = static("lecture/" + lecture_dir[question.lecture_name] + file_name)
+    if request.user.is_authenticated:
+        question = Question.objects.get(pk=id)
 
-    context = {"question": question, "src_text": src_text}
-    return render(request, "view/question_detail.html", context)
+        file_name = "/Page" + str(question.lecture_slide) + ".jpeg"
+        src_text = static("lecture/" + lecture_dir[question.lecture_name] + file_name)
+
+        context = {"question": question, "src_text": src_text}
+        return render(request, "view/question_detail.html", context)
+    else:
+        return redirect("/question/signin/")
 
 def question_delete(request, id):
-    question = Question.objects.get(pk=id)
-    question.delete()
+    if request.user.is_authenticated:
+        question = Question.objects.get(pk=id)
+        question.delete()
 
-    return redirect("/question/")
+        return redirect("/question/")
+    else:
+        return redirect("/question/signin/")
 
 def load_pages(request):
-    lecture_id = request.GET.get('lecture_id')
-    pages = []
-    if (lecture_id != ""):
-        pages = range(1, lectures[lecture_id]+1)
-    return render(request, 'component/page_dropdown_options.html', {'pages': pages})
+    if request.user.is_authenticated:
+        lecture_id = request.GET.get('lecture_id')
+        pages = []
+        if (lecture_id != ""):
+            pages = range(1, lectures[lecture_id]+1)
+        return render(request, 'component/page_dropdown_options.html', {'pages': pages})
+    else:
+        return redirect("/question/signin/")
 
 def load_slide(request):
     lecture = request.GET.get('lecture')
